Создал папку для кривых
