+++ TagsCloudVisualization/TagsCloudVisualization/TagsCloudVisualization/Curves/ICurve.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Curves
 {
 	public interface ICurve
 	{
