+++ TagsCloudVisualization/TagsCloudVisualization/TagsCloudVisualization/Curves/Spiral.cs
@@ -1,16 +1,16 @@
 ﻿using System;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Curves
 {
-	class Astroid : ICurve
+	public class Spiral : ICurve
 	{
 		private readonly double degreeStep;
 		private readonly double factorStep;
 		private int nextPointCounter;
 		private Point center;
 
-		public Astroid(double factorStep, double degreeStep, Point center)
+		public Spiral(double factorStep, double degreeStep, Point center)
 		{
 			this.factorStep = factorStep;
 			this.degreeStep = degreeStep;
@@ -22,8 +22,8 @@ namespace TagsCloudVisualization
 		{
 			var degree = degreeStep * nextPointCounter;
 			var factor = factorStep * nextPointCounter;
-			var x = center.X + (int)(factor * Math.Pow(Math.Cos(degree), 3));
-			var y = center.Y + (int)(factor * Math.Pow(Math.Sin(degree), 3));
+			var x = center.X + (int)(factor * Math.Sin(degree));
+			var y = center.Y + (int)(factor * Math.Cos(degree));
 			nextPointCounter++;
 
 			return new Point(x, y);
