+++ TagsCloudVisualization/TagsCloudVisualization/TagsCloudVisualization/Program.cs
@@ -2,6 +2,7 @@ using System;
 using System.ComponentModel;
 using System.Drawing;
 using System.Windows.Forms;
+using TagsCloudVisualization.Curves;
 
 namespace TagsCloudVisualization
 {
